+++ splayout/algorithms/__init__.py
@@ -1,5 +1,5 @@
 from .binarybatalgorithm import BinaryBatAlgorithm
 from .directbinarysearchalgorithm import DirectBinarySearchAlgorithm
-from .particleswarmalgorithm import ParitcleSwarmAlgorithm
-from .binaryparticleswarmalgorithm import BinaryParitcleSwarmAlgorithm
+from .particleswarmalgorithm import ParticleSwarmAlgorithm
+from .binaryparticleswarmalgorithm import BinaryParticleSwarmAlgorithm
 from .binarygeneticalgorithm import BinaryGeneticAlgorithm